+++ packages/cx/src/charts/helpers/ValueAtFinder.d.ts
@@ -1,16 +1,15 @@
-import * as Cx from '../../core';
-import {PointReducerProps} from './PointReducer';
+import * as Cx from "../../core";
+import { PointReducerProps } from "./PointReducer";
 
 interface ValueAtFinderProps extends PointReducerProps {
+   /* X axis probe value. */
+   at?: Cx.NumberProp | Cx.StringProp;
 
-   /* Probe value. */
-   at?: Cx.NumberProp,
-
-   /* A binding used to receive the measured value */
-   value?: Cx.Bind,
+   /* A binding used to receive the measured y axis value */
+   value?: Cx.Bind | Cx.AccessorChain<number>;
 
    /* A function used to convert x values into numeric format. Commonly used with dates. */
-   convert?: (value: any) => number
+   convert?: (value: number | string) => number;
 }
 
 /** Calculate value at a given point on the graph */